requirements.txt
aiofiles==0.7.0
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
fastapi==0.68.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
h11==0.12.0
httptools==0.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
pydantic==1.8.2
python-dotenv==0.19.0
python-jose==3.2.0
PyYAML==5.4.1
regex==2021.8.3
requests-oauthlib==1.3.0
starlette==0.14.2
tomli==1.2.0
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1